backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
boto3
opencv-python-headless
python-dotenv
